:sparkles: feat: 테스트 실행을 위한 byte 자료형을 int 로 변경 #102
- h2 tinyint 변환 오류

+++ src/main/java/com/habitpay/habitpay/domain/challenge/domain/Challenge.java
@@ -43,7 +43,7 @@ public class Challenge extends BaseTime {
     private int numberOfParticipants;
 
     @Column(nullable = false)
-    private byte participatingDays;
+    private int participatingDays;
 
     @Column(nullable = false)
     private int feePerAbsence;
